backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
razorpay==1.4.2
sqlparse==0.5.5
tzdata==2026.1
